+++ CHANGELOG.md
@@ -2,8 +2,60 @@
 
 ## Unreleased
 
-* Update the Go compiler from v1.25.5 to v1.25.7 ([#4383](https://github.com/evanw/esbuild/issues/4383)
+* Preserve URL fragments in data URLs ([#4370](https://github.com/evanw/esbuild/issues/4370))
 
+    Consider the following HTML, CSS, and SVG:
+
+    * `index.html`:
+
+        ```html
+        <!DOCTYPE html>
+        <html>
+          <head><link rel="stylesheet" href="icons.css"></head>
+          <body><div class="triangle"></div></body>
+        </html>
+        ```
+
+    * `icons.css`:
+
+        ```css
+        .triangle {
+          width: 10px;
+          height: 10px;
+          background: currentColor;
+          clip-path: url(./triangle.svg#x);
+        }
+        ```
+
+    * `triangle.svg`:
+
+        ```xml
+        <svg xmlns="http://www.w3.org/2000/svg">
+          <defs>
+            <clipPath id="x">
+              <path d="M0 0H10V10Z"/>
+            </clipPath>
+          </defs>
+        </svg>
+        ```
+
+    The CSS uses a URL fragment (the `#x`) to reference the `clipPath` element in the SVG file. Previously esbuild's CSS bundler didn't preserve the URL fragment when bundling the SVG using the `dataurl` loader, which broke the bundled CSS. With this release, esbuild will now preserve the URL fragment in the bundled CSS:
+
+    ```css
+    /* icons.css */
+    .triangle {
+      width: 10px;
+      height: 10px;
+      background: currentColor;
+      clip-path: url('data:image/svg+xml,<svg xmlns="http://www.w3.org/2000/svg"><defs><clipPath id="x"><path d="M0 0H10V10Z"/></clipPath></defs></svg>#x');
+    }
+    ```
+
+* Fix a minification bug with lowering of `for await` ([#4378](https://github.com/evanw/esbuild/pull/4378), [#4385](https://github.com/evanw/esbuild/pull/4385))
+
+    This release fixes a bug where the minifier would incorrectly strip the variable in the automatically-generated `catch` clause of lowered `for await` loops. The code that generated the loop previously failed to mark the internal variable references as used.
+
+* Update the Go compiler from v1.25.5 to v1.25.7 ([#4383](https://github.com/evanw/esbuild/issues/4383)
 
 ## 0.27.2